+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Fix stdin being closed by piping streams into it
 * Fix telnet echoLines option
 * Create new vlc instance when playing another stream
+* Add "load" method, to start loading the stream without playing
 
 ## 0.1.0 (2016-01-08)